use single unit test command to run tests for scratch/benchmark
